+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 **FaaS**
 - Add `pull` and `tag` parameters to `build` method.
+- Add `gateway` and `namespace` parameters to `deploy` method.
 
 **GoSS**
 - Add `logLevel` parameter to `server` and `validate` methods.